Phase 3 (suite): resoudre C4018/C4245/C4389 sous /W4
Incompatibilites signe/non-signe (comparaisons de boucles et
sentinelles -1), corrigees au cas par cas :

- La plupart : boucles ascendantes depuis 0 dont la variable etait
  declaree int alors que la borne comparee est unsigned int -- type
  de la variable change en unsigned int quand elle n'est utilisee que
  dans cette boucle (pas de risque de nouveau warning ailleurs).
- Object3DLWOResource.cpp : jj est reutilisee dans 3 boucles du meme
  bloc comparant a des champs de signes differents (m_MaxNbTriangle
  est signe, MaxNbVertex/m_NumberOfSurfaceUsed ne le sont pas) --
  garde en int, cast ponctuel la ou necessaire plutot que de changer
  son type (aurait invente un nouveau warning sur la boucle deja
  saine).
- VedaLib3DEngine/lwo2reader.c : jj/kk/numvv sont de vraies variables
  signees utilisees comme sentinelles -1 ("non trouve") alors que
  declarees unsigned int -- le code fonctionnait par hasard grace au
  rebouclage implicite unsigned, mais c'est fragile ; separees en
  declarations int propres.
- VedaMachineOGL/OGLMachine.cpp : m_flags(-1) dans un initialiseur de
  constructeur (idiome "tous les bits a 1") remplace par m_flags(~0u),
  explicite et sans avertissement, valeur identique.
- Reste (casts ponctuels a la comparaison, sans changer de type) :
  Object3DScene.cpp, Image2DEffectFire.cpp, Object3DMarchCube.cpp,
  VirtualImage.cpp, EquationSoundSignal.cpp, MP3SoundObject.cpp,
  BaseObject.cpp, TextLocalized.cpp.

Warnings AzurVeda restants sous /W4 : 209 -> 180.

Build revalide (exit 0, aucune erreur).

+++ Veda/TextLocalized.cpp
@@ -50,7 +50,7 @@ void	TextLocalized::SynchronizeTextToMachineLanguage()
 	while(pCell)
 	{
 		LocalText *pLocalText = (LocalText *) pCell->GetManagedObject();
-		if( pLocalText->mSer_LocalizationEnum.Get() == currentLanguage )
+		if( (int)pLocalText->mSer_LocalizationEnum.Get() == currentLanguage )
 		{
 			m_pCurrentText = &(pLocalText->mSer_BaseText);
 			return;
